+++ components/analytics/shake2report/impl/src/main/java/com/flipperdevices/analytics/shake2report/impl/InternalShake2Report.kt
@@ -1,5 +1,6 @@
 package com.flipperdevices.analytics.shake2report.impl
 
+import kotlinx.coroutines.flow.StateFlow
 import java.io.File
 
 interface InternalShake2Report {
@@ -7,4 +8,5 @@ interface InternalShake2Report {
 
     fun register()
     fun setExtra(tags: List<Pair<String, String>>)
+    fun getIsRegisteredFlow(): StateFlow<Boolean>
 }