+++ components/analytics/shake2report/impl/src/main/java/com/flipperdevices/analytics/shake2report/impl/api/Shake2ReportApiImpl.kt
@@ -32,6 +32,8 @@ class Shake2ReportApiImpl @Inject constructor(
 
         Sentry.captureEvent(event)
     }
+
+    override fun isInitialized() = internalShake2Report.getIsRegisteredFlow()
 }
 
 private const val FLIPPER_DEVICE_NAME = "flipper_device_name"