+++ components/analytics/shake2report/noop/src/main/java/com/flipperdevices/shake2report/noop/Shake2ReportStub.kt
@@ -4,6 +4,7 @@ import com.flipperdevices.bridge.api.model.FlipperGATTInformation
 import com.flipperdevices.core.di.AppGraph
 import com.flipperdevices.shake2report.api.Shake2ReportApi
 import com.squareup.anvil.annotations.ContributesBinding
+import kotlinx.coroutines.flow.MutableStateFlow
 import javax.inject.Inject
 
 @ContributesBinding(AppGraph::class)
@@ -22,4 +23,6 @@ class Shake2ReportStub @Inject constructor() : Shake2ReportApi {
     override fun reportException(throwable: Throwable, tag: String?, extras: Map<String, String>?) {
         // Do nothing
     }
+
+    override fun isInitialized() = MutableStateFlow(true)
 }