+++ components/settings/impl/src/main/java/com/flipperdevices/settings/impl/viewmodels/SettingsViewModel.kt
@@ -11,8 +11,10 @@ import com.flipperdevices.core.log.error
 import com.flipperdevices.core.preference.pb.SelectedTheme
 import com.flipperdevices.core.preference.pb.Settings
 import com.flipperdevices.core.share.ShareHelper
+import com.flipperdevices.selfupdater.api.SelfUpdaterApi
 import com.flipperdevices.settings.impl.R
 import com.flipperdevices.settings.impl.model.ExportState
+import com.flipperdevices.shake2report.api.Shake2ReportApi
 import com.flipperdevices.shake2report.api.Shake2ReportFeatureEntry
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.MutableStateFlow
@@ -23,12 +25,14 @@ import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
 import tangle.viewmodel.VMInject
 
-@Suppress("LongParameterList")
+@Suppress("LongParameterList", "TooManyFunctions")
 class SettingsViewModel @VMInject constructor(
     private val dataStoreSettings: DataStore<Settings>,
     private val applicationParams: ApplicationParams,
     private val exportKeysHelper: ExportKeysHelper,
-    private val shake2ReportFeatureEntry: Shake2ReportFeatureEntry
+    private val shake2ReportFeatureEntry: Shake2ReportFeatureEntry,
+    private val shake2ReportApi: Shake2ReportApi,
+    private val selfUpdaterApi: SelfUpdaterApi
 ) : ViewModel(), LogTagProvider {
     override val TAG = "SettingsViewModel"
 
@@ -44,6 +48,8 @@ class SettingsViewModel @VMInject constructor(
     fun getState(): StateFlow<Settings> = settingsState
 
     fun getExportState(): StateFlow<ExportState> = exportStateFlow
+
+    fun getShake2ReportInitializationState(): StateFlow<Boolean> = shake2ReportApi.isInitialized()
 
     fun onSwitchDebug(value: Boolean) {
         viewModelScope.launch {
@@ -112,4 +118,5 @@ class SettingsViewModel @VMInject constructor(
     }
 
     fun versionApp() = applicationParams.version
+    fun sourceInstall() = selfUpdaterApi.getInstallSourceName()
 }