+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - [CI] Migrate to Detekt
 - [CI] Android lint in CI
 - [CI] Enable back github queue
+- [CI] Migrate to matrix CI with source install
 - [FIX] Clickable modifier ktx
 - [FIX] Refactor `share:receive` module
 - [FIX] Bump length for server-side share link